Refactored ADR code for PR#2175
