+++ manage.py
@@ -37,7 +37,7 @@ manager.command(legal_docs.index_statutes)
 manager.command(legal_docs.load_advisory_opinions_into_s3)
 manager.command(legal_docs.load_archived_murs)
 manager.command(legal_docs.load_current_murs)
-manager.command(legal_docs.load_current_adrs)
+manager.command(legal_docs.load_adrs)
 manager.command(legal_docs.initialize_legal_docs)
 manager.command(legal_docs.create_staging_index)
 manager.command(legal_docs.restore_from_staging_index)
