+++ webservices/legal_docs/utils.py
@@ -1,47 +1,41 @@
 import logging
 import re
+import pytest
 from collections import defaultdict
 from urllib.parse import urlencode
 
 from webservices.env import env
 from webservices.legal_docs import DOCS_INDEX
+from webservices.utils import create_eregs_link
 from webservices.rest import db
-from webservices.utils import create_eregs_link, get_elasticsearch_connection
-from webservices.tasks.utils import get_bucket
 
 from .reclassify_statutory_citation import reclassify_current_mur_statutory_citation
 
 logger = logging.getLogger(__name__)
 
-ALL_ADRS = """
-    SELECT case_id, case_no, name
-    FROM fecmur.case
-    WHERE case_type = 'ADR'
-"""
+STATUTE_REGEX = re.compile(r'(?<!\(|\d)(?P<section>\d+([a-z](-1)?)?)')
+REGULATION_REGEX = re.compile(r'(?<!\()(?P<part>\d+)(\.(?P<section>\d+))?')
 
-ADR_SUBJECTS = """
+SUBJECTS = """
     SELECT subject.description AS subj, relatedsubject.description AS rel
     FROM fecmur.case_subject
     JOIN fecmur.subject USING (subject_id)
     LEFT OUTER JOIN fecmur.relatedsubject USING (subject_id, relatedsubject_id)
     WHERE case_id = %s
 """
-
-ADR_ELECTION_CYCLES = """
+ELECTION_CYCLES = """
     SELECT election_cycle::INT
     FROM fecmur.electioncycle
     WHERE case_id = %s
 """
-
-ADR_PARTICIPANTS = """
+PARTICIPANTS = """
     SELECT entity_id, name, role.description AS role
     FROM fecmur.players
     JOIN fecmur.role USING (role_id)
     JOIN fecmur.entity USING (entity_id)
     WHERE case_id = %s
 """
-
-ADR_DOCUMENTS = """
+DOCUMENTS = """
     SELECT document_id, category, description, ocrtext,
         fileimage, length(fileimage) AS length,
         doc_order_id, document_date
@@ -51,19 +45,17 @@ ADR_DOCUMENTS = """
 """
 # TODO: Check if document order matters
 
-ADR_VIOLATIONS = """
+VIOLATIONS = """
     SELECT entity_id, stage, statutory_citation, regulatory_citation
     FROM fecmur.violations
     WHERE case_id = %s
     ;
 """
-
 OPEN_AND_CLOSE_DATES = """
     SELECT min(event_date), max(event_date)
     FROM fecmur.calendar
     WHERE case_id = %s;
 """
-
 DISPOSITION_DATA = """
     SELECT fecmur.event.event_name,
     fecmur.settlement.final_amount, fecmur.entity.name, violations.statutory_citation,
@@ -79,62 +71,30 @@ DISPOSITION_DATA = """
     where fecmur.calendar.case_id={0} and event_name not in ('Complaint/Referral', 'Disposition')
     ORDER BY fecmur.event.event_name ASC, fecmur.settlement.final_amount DESC NULLS LAST, event_date DESC;
 """
-
 DISPOSITION_TEXT = """
 SELECT vote_date, action from fecmur.commission
 WHERE case_id = %s
 ORDER BY vote_date desc;
 """
-
-STATUTE_REGEX = re.compile(r'(?<!\(|\d)(?P<section>\d+([a-z](-1)?)?)')
-REGULATION_REGEX = re.compile(r'(?<!\()(?P<part>\d+)(\.(?P<section>\d+))?')
-
-def load_current_adrs():
-    """
-    Reads data for current MURs from a Postgres database, assembles a JSON document
-    corresponding to the MUR and indexes this document in Elasticsearch in the index
-    `docs_index` with a doc_type of `murs`. In addition, all documents attached to
-    the MUR are uploaded to an S3 bucket under the _directory_ `legal/murs/current/`.
-    """
-    es = get_elasticsearch_connection()
-    bucket = get_bucket()
-    bucket_name = env.get_credential('bucket')
+def get_subjects(case_id):
+    subjects = []
     with db.engine.connect() as conn:
-        rs = conn.execute(ALL_ADRS)
+        rs = conn.execute(SUBJECTS, case_id)
         for row in rs:
-            case_id = row['case_id']
-            adr = {
-                'doc_id': 'adr_%s' % row['case_no'],
-                'no': row['case_no'],
-                'name': row['name'],
-                'adr_type': 'current',
-            }
-            adr['subject'] = {"text": get_subjects(case_id)}
-            adr['election_cycles'] = get_election_cycles(case_id)
-
-            participants = get_participants(case_id)
-            adr['participants'] = list(participants.values())
-            adr['respondents'] = get_sorted_respondents(adr['participants'])
-            adr['disposition'] = get_disposition(case_id)
-            adr['documents'] = get_documents(case_id, bucket, bucket_name)
-            adr['open_date'], adr['close_date'] = get_open_and_close_dates(case_id)
-            adr['url'] = '/legal/matter-under-review/%s/' % row['case_no']
-            es.index(DOCS_INDEX, 'adrs', adr, id=adr['doc_id'])
-            print('Inside curret_adrs()...., adr : ',adr)
+            if row['rel']:
+                subject_str = row['subj'] + "-" + row['rel']
+            else:
+                subject_str = row['subj']
+            subjects.append(subject_str)
+    return subjects
 
 def get_election_cycles(case_id):
     election_cycles = []
     with db.engine.connect() as conn:
-        rs = conn.execute(ADR_ELECTION_CYCLES, case_id)
+        rs = conn.execute(ELECTION_CYCLES, case_id)
         for row in rs:
             election_cycles.append(row['election_cycle'])
     return election_cycles
-
-def get_open_and_close_dates(case_id):
-    with db.engine.connect() as conn:
-        rs = conn.execute(OPEN_AND_CLOSE_DATES, case_id)
-        open_date, close_date = rs.fetchone()
-    return open_date, close_date
 
 def get_disposition(case_id):
     with db.engine.connect() as conn:
@@ -152,10 +112,11 @@ def get_disposition(case_id):
             disposition_text.append({'vote_date': row['vote_date'], 'text': row['action']})
         return {'text': disposition_text, 'data': disposition_data}
 
+
 def get_participants(case_id):
     participants = {}
     with db.engine.connect() as conn:
-        rs = conn.execute(ADR_PARTICIPANTS, case_id)
+        rs = conn.execute(PARTICIPANTS, case_id)
         for row in rs:
             participants[row['entity_id']] = {
                 'name': row['name'],
@@ -164,31 +125,30 @@ def get_participants(case_id):
             }
     return participants
 
-def get_sorted_respondents(participants):
-    """
-    Returns the respondents in a MUR sorted in the order of most important to least important
-    """
-    SORTED_RESPONDENT_ROLES = ['Primary Respondent', 'Respondent', 'Previous Respondent']
-    respondents = []
-    for role in SORTED_RESPONDENT_ROLES:
-        respondents.extend(sorted([p['name'] for p in participants if p['role'] == role]))
-    return respondents
-
-def get_subjects(case_id):
-    subjects = []
+def get_documents(case_id, bucket, bucket_name):
+    documents = []
     with db.engine.connect() as conn:
-        rs = conn.execute(ADR_SUBJECTS, case_id)
+        rs = conn.execute(DOCUMENTS, case_id)
         for row in rs:
-            if row['rel']:
-                subject_str = row['subj'] + "-" + row['rel']
-            else:
-                subject_str = row['subj']
-            subjects.append(subject_str)
-    return subjects
+            document = {
+                'document_id': row['document_id'],
+                'category': row['category'],
+                'description': row['description'],
+                'length': row['length'],
+                'text': row['ocrtext'],
+                'document_date': row['document_date'],
+            }
+            pdf_key = 'legal/murs/current/%s.pdf' % row['document_id']
+            logger.info("S3: Uploading {}".format(pdf_key))
+            bucket.put_object(Key=pdf_key, Body=bytes(row['fileimage']),
+                    ContentType='application/pdf', ACL='public-read')
+            document['url'] = "https://%s.s3.amazonaws.com/%s" % (bucket_name, pdf_key)
+            documents.append(document)
+    return documents
 
 def assign_citations(participants, case_id):
     with db.engine.connect() as conn:
-        rs = conn.execute(ADR_VIOLATIONS, case_id)
+        rs = conn.execute(VIOLATIONS, case_id)
         for row in rs:
             entity_id = row['entity_id']
             if entity_id not in participants:
@@ -198,7 +158,7 @@ def assign_citations(participants, case_id):
                 parse_statutory_citations(row['statutory_citation'], case_id, entity_id))
             participants[entity_id]['citations'][row['stage']].extend(
                 parse_regulatory_citations(row['regulatory_citation'], case_id, entity_id))
-
+            
 def parse_statutory_citations(statutory_citation, case_id, entity_id):
     citations = []
     if statutory_citation:
@@ -245,26 +205,21 @@ def parse_regulatory_citations(regulatory_citation, case_id, entity_id):
                     regulatory_citation, entity_id, case_id)
     return citations
 
-def get_documents(case_id, bucket, bucket_name):
-    documents = []
+def get_open_and_close_dates(case_id):
     with db.engine.connect() as conn:
-        rs = conn.execute(ADR_DOCUMENTS, case_id)
-        for row in rs:
-            document = {
-                'document_id': row['document_id'],
-                'category': row['category'],
-                'description': row['description'],
-                'length': row['length'],
-                'text': row['ocrtext'],
-                'document_date': row['document_date'],
-            }
-            pdf_key = 'legal/murs/current/%s.pdf' % row['document_id']
-            logger.info("S3: Uploading {}".format(pdf_key))
-            bucket.put_object(Key=pdf_key, Body=bytes(row['fileimage']),
-                    ContentType='application/pdf', ACL='public-read')
-            document['url'] = "https://%s.s3.amazonaws.com/%s" % (bucket_name, pdf_key)
-            documents.append(document)
-    return documents
+        rs = conn.execute(OPEN_AND_CLOSE_DATES, case_id)
+        open_date, close_date = rs.fetchone()
+    return open_date, close_date
+
+def get_sorted_respondents(participants):
+    """
+    Returns the respondents in a MUR sorted in the order of most important to least important
+    """
+    SORTED_RESPONDENT_ROLES = ['Primary Respondent', 'Respondent', 'Previous Respondent']
+    respondents = []
+    for role in SORTED_RESPONDENT_ROLES:
+        respondents.extend(sorted([p['name'] for p in participants if p['role'] == role]))
+    return respondents
 
 def remove_reclassification_notes(statutory_citation):
     """ Statutory citations include notes on reclassification of the form
